mysql/json: move parseNumberType to the test that is its last reader
Settling a number's kind at parse time left parseNumberType with no
production caller. The one reader left is the test that checks the
shape rule against the conversions, where it serves as the oracle, so
it lives with that test.
